Refactor css
Created a global css file that is used for all (ok, both) pages in the project, and individual css files that are for each page. I did this to kepe me from having to mess with or move properties that are important to index.html while building & designing results-page.html, and to keep each individual css file a little easier to deal with.